+++ src/data_parser.py
@@ -1,6 +1,9 @@
 import json
 import os
+import random
 import re
+
+from tqdm import tqdm
 
 
 def save_tasks(tasks_dict, task_file):
@@ -14,11 +17,11 @@ def save_tasks(tasks_dict, task_file):
     if os.path.exists(task_file):
         with open(task_file, 'a', errors='ignore') as file:
             for signature, body in tasks_dict.items():
-                file.write('{"signature": "' + signature + '", "body": "' + body + '"}\n')
+                file.write(json.dumps({"signature": signature, "body": body}) + '\n')
     else:
         with open(task_file, 'w', errors='ignore') as file:
             for signature, body in tasks_dict.items():
-                file.write('{"signature": "' + signature + '", "body": "' + body + '"}\n')
+                file.write(json.dumps({"signature": signature, "body": body}) + '\n')
 
 
 def split_signature_body(matches, split_char):
@@ -67,6 +70,9 @@ def filter_invalid(match):
     if match.count('DEBUG'):
         return None
 
+    if match.count('FIXME'):
+        return None
+
     return match
 
 
@@ -88,8 +94,16 @@ def format_match(match):
     for comment in multiline_comments:
         match = match.replace(comment, '')
 
-    match = match.replace('\n', '\\n')
-    match = match.replace('    ', '')
+    string_literals = re.findall(r'".*?"', match)
+    for string_literal in string_literals:
+        match = match.replace(string_literal, '<STR_LIT>')
+
+    number_literals = re.findall(r'\d+', match)
+    for number_literal in number_literals:
+        match = match.replace(number_literal, '<NUM_LIT>')
+
+    match = match.replace('\n', '<EOL>')
+    match = match.replace('    ', '<INDENT>')
 
     return match
 
@@ -118,17 +132,7 @@ def find_matches(code, patterns):
     return all_matches
 
 
-def extract_completion_tasks(code):
-    """
-    Extract completion tasks from the provided code
-
-    :param code:
-    :return:
-    """
-    data_dir = "../data/Kotlin/completion_tasks/"
-    if not os.path.exists(data_dir):
-        os.makedirs(data_dir)
-
+def extract_completion_tasks(data_dir, output_dir):
     task_files = ["condition_tasks.jsonl",
                   "function_tasks.jsonl",
                   "import_tasks.jsonl",
@@ -154,22 +158,35 @@ def extract_completion_tasks(code):
 
     split_chars = ['{', '{', '.', '{', '=']
 
-    for task_file, task_pattern, split_char in zip(task_files, task_patterns, split_chars):
-        matches = find_matches(code, task_pattern)
-        tasks_dict = split_signature_body(matches, split_char)
-        save_tasks(tasks_dict, data_dir + task_file)
+    kotlin_files = os.listdir(data_dir)
+
+    for task_file, task_pattern, split_char in tqdm(zip(task_files, task_patterns, split_chars), desc="Fetching tasks"):
+        found_tasks = {}
+        for file in kotlin_files:
+            with open(data_dir + file, 'r', encoding='utf-8') as f:
+                content = f.read()
+
+            matches = find_matches(content, task_pattern)
+            tasks_dict = split_signature_body(matches, split_char)
+
+            found_tasks.update(tasks_dict)
+
+        save_tasks(found_tasks, output_dir + task_file)
+
+    tasks_summary(output_dir)
 
 
-def tasks_summary():
+def tasks_summary(data_dir):
     """
     Print the summary of completion tasks
 
     :return:
     """
-    data_dir = "../data/Kotlin/completion_tasks/"
-
     total_tasks = 0
     for task_file in os.listdir(data_dir):
+        if os.path.isdir(data_dir + task_file):
+            continue
+
         with open(data_dir + task_file, 'r') as file:
             lines = file.readlines()
             print(task_file + ': ' + str(len(lines)))
@@ -177,6 +194,30 @@ def tasks_summary():
 
     print('--------------------')
     print('Total number: ' + str(total_tasks))
+
+
+def split_data(data_dir, output_dir, split_ratio=0.8):
+    combined_data = []
+    for task_file in os.listdir(data_dir):
+        with open(data_dir + task_file, 'r') as file:
+            data = [json.loads(line) for line in file]
+            combined_data.extend(data)
+
+    random.shuffle(combined_data)
+
+    split_index = int(split_ratio * len(combined_data))
+    train_data = combined_data[:split_index]
+    test_data = combined_data[split_index:]
+
+    with open(output_dir + 'train.jsonl', 'w') as file:
+        for item in train_data:
+            file.write(json.dumps(item) + '\n')
+
+    with open(output_dir + 'test.jsonl', 'w') as file:
+        for item in test_data:
+            file.write(json.dumps(item) + '\n')
+
+    tasks_summary(output_dir)
 
 
 def load_dataset(file_path, batch_size):